WIP - not working. About to attempt what Tridge suggested via mumble
Make the structures smaller and contain less infomation.

Two structures with a single uint_8 pointing to the current structure
(the one that is not currently being written to).

The read time of the structure will be minimal, dont worry about it if
the camera is operating at 30Hz.

Make the structures static and put a block comment above it saying why
it is static (accessed from a callback function within AP_Camera.

Keep the headder files clean and free from code. code in the headder
file will increase flash size which in gerneral impacts ardupilot more
than the speed tradeoff.