WIP - still not working but building with static AHRS and Camera members
Now the camera data in AHRS is static and can be accessed from the
callback.

The old issue of the camera trigger event linking to the AP_Camera
callback (and not the AP_Camera_Vision callback is still an issue).

Need to drop AP_Camera_Vision and modify AP_Camera directly